pyproject.toml
[project]
name = "diffusionsr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "basicsr>=1.4.2",
    "einops>=0.7.0",
    "ipykernel>=7.1.0",
    "loralib>=0.1.2",
    "lpips>=0.1.4",
    "matplotlib>=3.10.7",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "wandb>=0.23.0",
]
